chore: More reliable definition of C standard in CMakeLists (#605)
Some local update on my local machine resulted in nanoarrow not
compiling because clang gave a warning for the C11 extensions in the
flatcc generated code. I believe this was because the C standard had
already been set to C99 at the beginning of the file, whereas the check
to update it to C11 only happened after this had already been set.