chore(python): Fix type stubs for updated module structure (#606)
This PR updates the type stubs for the Cython modules to reflect the
updated structure. I don't think this mechanism for generating stubs is
the best long-term strategy (requires remembering to do so!)...we could
also do something like omit the generated stubs from the source tree and
generate them only before packaging; however, this would make them not
useful for local development (which is one of the primary short-term
benefits of including them).