chore(python): Update Python packaging (#625)
This PR updates the Python packaging to build a single nanoarrow library
that is shared amongst the various extension modules and updates the
version of cibuildwheel such that we can release Python 3.13 wheels in
our upcoming release. I'm not sure where the mechanism used to build a
"shared" library is documented, nor am I sure that it is truly a
"shared" library (may be static); however, it seems conceptually better
than the existing system of compiling the source files more than once
for each module that needed them.