feat: Add ArrowArrayView accessors to inspect buffer properties (#638)
This PR abstracts accessors for the buffer_view, buffer type, buffer
data type, and element bit width for the `ArrowArrayView`. Before adding
string/binary view support, this was done by directly accessing the
`layout` and `buffer_view` members; however, this required
special-casing + some duplicated code in the string view in the R/Python
bindings.

This PR also removes the dependence on the `ArrowArrayView::array`
member, since this member is optional (i.e., the data backing an
`ArrowArrayView` need not be related to an actual `ArrowArray`).