feat: Add ZSTD decompression support to IPC reader (#693)
This PR implements ZSTD buffer decompression in the `ArrowIpcDecoder`
and in the `ArrowArrayStreamReader` when built with
`-DNANOARROW_IPC_WITH_ZSTD=ON`. It also allows a user to inject support
for these into the `ArrowIpcDecoder` if for whatever reason they don't
have control over the build flags (or want to use ZSTD that has been
made available to them in a different way).

This doesn't implement multithreaded decompression but does allow a user
to implement it by not using the default `ArrowIpcSerialDecompressor()`.
This could be included in header-only C++ if there is some interest.

A non-trivial example in Python bindings (where were also wired up to
support it):

```python
import urllib.request
import nanoarrow as na
url = "https://github.com/geoarrow/geoarrow-data/releases/download/v0.1.0/ns-water-basin_point-wkb.arrow"

# Work around the 'no arrow file support'
with urllib.request.urlopen(url) as f:
    f.read(8)
    print(na.ArrayStream.from_readable(f).read_all())
#> nanoarrow.Array<non-nullable struct<OBJECTID: int64, FEAT_CODE: string, ...>[46]
#> {'OBJECTID': 1, 'FEAT_CODE': 'WABA30', 'BASIN_NAME': '01EB000', 'RIVER': 'BAR...
#> {'OBJECTID': 2, 'FEAT_CODE': 'WABA30', 'BASIN_NAME': '01EC000', 'RIVER': 'ROS...
#> {'OBJECTID': 3, 'FEAT_CODE': 'WABA30', 'BASIN_NAME': '01EA000', 'RIVER': 'TUS...
#> {'OBJECTID': 4, 'FEAT_CODE': 'WABA30', 'BASIN_NAME': '01DA000', 'RIVER': 'MET...
#> {'OBJECTID': 5, 'FEAT_CODE': 'WABA30', 'BASIN_NAME': '01ED000', 'RIVER': 'MER...
#> {'OBJECTID': 6, 'FEAT_CODE': 'WABA30', 'BASIN_NAME': '01EE000', 'RIVER': 'HER...
#> {'OBJECTID': 7, 'FEAT_CODE': 'WABA30', 'BASIN_NAME': '01EG000', 'RIVER': 'GOL...
#> {'OBJECTID': 8, 'FEAT_CODE': 'WABA30', 'BASIN_NAME': '01EF000', 'RIVER': 'LAH...
#> {'OBJECTID': 9, 'FEAT_CODE': 'WABA30', 'BASIN_NAME': '01EJ000', 'RIVER': 'SAC...
#> {'OBJECTID': 10, 'FEAT_CODE': 'WABA30', 'BASIN_NAME': '01EH000', 'RIVER': 'EA...
#> ...and 36 more items
```

This PR doesn't implement the R bindings (because adding a zstd
dependency there is a can of worms better suited to another PR).