chore(ci): Tweak compose, image, and ignores (#698)
This PR moves coverage back to an image-based compose service rather
than build-based one (which we can do now that #694 is merged and the
images are rebuilt and uploaded).

I also discovered that the alpine build was still using an old Arrow C++
version (which was causing occasional failures due to the low
availability of archive.apache.org). I also found a generated file
directory when running coverage locally.