chore: Fall back on shared libzstd if libzstd_static is not available (#704)
This PR updates the CMake such that it works in conda after a `conda
install zstd` (which doesn't provide `libzstd_static`, but does provide
`libzstd`). This is needed to unskip the integration tests for zstd
because the integration image uses conda (
https://github.com/apache/arrow/pull/45205 ).

Previous commits tested https://github.com/apache/arrow/pull/45205 by
specifically pulling that branch instead of apache/arrow. When zstd is
installed...the tests pass!