feat: Add `ArrowDecimalAppendStringToBuffer()` (#720)
We've had `ArrowDecimalAppendDigitsToBuffer()` (because it was required
for integration testing), but this only gets the underlying integer
representation (e.g., 1234), not the "logical" decimal representation
(e.g., 1.234). This logical representation gets returned in database
results (e.g., it's what ADBC does for postgres) and can be used to
support convert to double where an implementation can parse a string
(e.g., in R).

This was extracted out of
https://github.com/apache/arrow-nanoarrow/pull/717 , which is a fix for
an ADBC issue ( https://github.com/apache/arrow-adbc/issues/2508 ).