feat: Improve shared linkage support (#719)
This PR makes linking to the usual targets static linkage and adds a
paired `_shared` target for users that want to opt in to shared
nanoarrow linkage. This is intended to mirror what ADBC and Arrow do
(with separated `_shared` and `_static` targets).

This PR also adds the appropriate import/export attribute for Windows
(most of the changed lines in this PR).

While I had my Windows machine fired up I also fixed a few compiler
warnings, and I added a Windows and MacOS run for the `examples`
workflow.

Closes #495, closes #718