fix: ArrowDecimalSetDigits() on s390x + test fixes for big endian (#732)
This PR fixes Decimal32 and Decimal64 handling on big endian. The
previous code used a cast on the least significant word to convert
between 64-bit words and the 32 bit value; however, this is only correct
on big endian. This was exposed by #720, which in turn exposed that the
get/set digits behaviour was not tested for Decimal 32/64. This code was
updated to use a `memcpy()` instead of a cast for the 32-bit case. The
s390x workflow had been failing because of a segfaulting compiler for a
while and so I'd missed that the decimal 32/64 change had broken it 😬 .

There is also one change to get the Windows verification job passing
that I noticed when running the verify.yaml workflow to test big endian.

This can be checked locally with:

```shell
export NANOARROW_ARCH=s390x  
docker compose run --rm verify     
```

(and was also checked via the verify.yaml workflow)