fix(ci): Fix runner specification for c-device (#760)
This is currently failing with:

```
Error when evaluating 'runs-on' for job 'test-c-device'. .github/workflows/build-and-test-device.yaml (Line: 40, Col: 14): Unexpected value ''
```

...because we'd been relying on the "empty string means ubuntu-latest"
that was removed from GitHub actions at some point. There are a few
updates here as well to fix the device build since that job hadn't been
running for some time.