chore: Add options to CMake configuration for vcpkg build (#788)
The vcpkg checks for a dry run of the current state of the main branch (
https://github.com/microsoft/vcpkg/pull/46029 ) fail because (1)
`-Werror` is always set in debug mode and there's at least one platform
where we get some extra warnings and (2) we install shared libraries
when static linkage was requested. This PR adds options so we can turn
those off in the vcpkg config.